add test for data type
